Fix wrong argument type

+++ ros/src/data/packages/map_file/nodes/vector_map_loader/vector_map_loader.cpp
@@ -161,7 +161,7 @@ std::vector<WhiteLine> read_whiteline(const char* filename)
     ret[id].lid = std::stoi(tbl[i][1]);
     ret[id].width = std::stod(tbl[i][2]);
     ret[id].color = tbl[i][3].c_str()[0];
-    ret[id].type = std::stod(tbl[i][4]);
+    ret[id].type = std::stoi(tbl[i][4]);
     ret[id].linkid = std::stoi(tbl[i][5]);
   }
   return ret;
@@ -347,7 +347,7 @@ struct PointClass {
   double h;
   double bx;
   double ly;
-  double ref;
+  int ref;
   int mcode1;
   int mcode2;
   int mcode3;
@@ -367,10 +367,10 @@ std::vector<PointClass> read_pointclass(const char *filename)
     ret[id].h = std::stod(tbl[i][3]);
     ret[id].bx = std::stod(tbl[i][4]);
     ret[id].ly = std::stod(tbl[i][5]);
-    ret[id].ref = std::stod(tbl[i][6]);
+    ret[id].ref = std::stoi(tbl[i][6]);
     ret[id].mcode1 = std::stoi(tbl[i][7]);
-    ret[id].mcode2 = std::stod(tbl[i][8]);
-    ret[id].mcode3 = std::stod(tbl[i][9]);
+    ret[id].mcode2 = std::stoi(tbl[i][8]);
+    ret[id].mcode3 = std::stoi(tbl[i][9]);
   }
   return ret;
 }
@@ -441,8 +441,8 @@ std::vector<AreaClass> read_areaclass(const char *filename)
   for (i=0; i<n; i++) {
     int id = std::stoi(tbl[i][0]);
     ret[id].aid = id;
-    ret[id].slid = std::stod(tbl[i][1]);
-    ret[id].elid = std::stod(tbl[i][2]);
+    ret[id].slid = std::stoi(tbl[i][1]);
+    ret[id].elid = std::stoi(tbl[i][2]);
   }
   return ret;
 }
@@ -507,7 +507,7 @@ struct DTLane {
   double dist;
   int pid;
   double dir;
-  int apara;
+  double apara;
   double r;
   double slope;
   double cant;
@@ -527,7 +527,7 @@ std::vector<DTLane> read_dtlane(const char *filename)
     ret[id].dist = std::stod(tbl[i][1]);
     ret[id].pid = std::stoi(tbl[i][2]);
     ret[id].dir = std::stod(tbl[i][3]);
-    ret[id].apara = std::stoi(tbl[i][4]);
+    ret[id].apara = std::stod(tbl[i][4]);
     ret[id].r = std::stod(tbl[i][5]);
     ret[id].slope = std::stod(tbl[i][6]);
     ret[id].cant = std::stod(tbl[i][7]);
